feat: Refactor login and email verification components; streamline code structure and improve user experience by using react suspense
